Deepen Flow launch coordinator review fixes
Move manual target selection and auto-launch coordination fully into the Flow launch module so key handling only delegates the launch action.

Normalize phase IDs for plan-review preflight, autoreview recovery, and prompt template selection, and add internal coordinator coverage for direct auto launch, suppression, deferred relaunch, and normalized phase IDs.
