+++ model/model_keys.go
@@ -1310,218 +1310,6 @@ func (m Model) handleLaunchNextFlowPhase() (tea.Model, tea.Cmd) {
 	return next.launchFlowPhaseTarget(target)
 }
 
-func (m Model) launchFlowPhaseTarget(target flowPhaseLaunchTarget) (tea.Model, tea.Cmd) {
-	return m, m.prepareFlowPhaseLaunch(target)
-}
-
-func (m Model) prepareAutoFlowPhaseLaunch(previousFlows, currentFlows []flowstore.FlowRecord) (Model, tea.Cmd) {
-	previousByFlowID := make(map[string]flowstore.FlowRecord, len(previousFlows))
-	for _, record := range previousFlows {
-		if record.FlowID != "" {
-			previousByFlowID[record.FlowID] = record
-		}
-	}
-	var cmds []tea.Cmd
-	for _, record := range currentFlows {
-		if !record.AutoMode || record.FlowID == "" {
-			continue
-		}
-		previous, ok := previousByFlowID[record.FlowID]
-		if !ok {
-			continue
-		}
-		completedPhase, ok := newlyCompletedFlowPhase(previous, record)
-		if !ok {
-			m = m.clearResolvedSuppressedAutoFlowLaunches(record)
-			continue
-		}
-		if artifacts.NormalizePhaseID(completedPhase.PhaseID) == "autoreview" {
-			continue
-		}
-		if m.isAutoFlowLaunchSuppressed(record.FlowID, completedPhase) {
-			m = m.clearSuppressedAutoFlowLaunch(record.FlowID, completedPhase)
-			continue
-		}
-		sourceLaunchID := flowstore.LatestPhaseLaunchID(completedPhase)
-		if m.hasRunningFlowEmbeddedTerminalForPhaseLaunch(record.FlowID, completedPhase.PhaseID, sourceLaunchID) ||
-			m.hasAutoClosingFlowEmbeddedTerminalForPhaseLaunch(record.FlowID, completedPhase.PhaseID, sourceLaunchID) {
-			m = m.deferAutoFlowPhaseLaunch(record.FlowID, completedPhase.PhaseID)
-			continue
-		}
-		if m.hasFlowEmbeddedTerminalForPhaseLaunch(record.FlowID, completedPhase.PhaseID, sourceLaunchID) {
-			m = m.suppressAutoFlowPhaseLaunch(record.FlowID, completedPhase.PhaseID, sourceLaunchID)
-			continue
-		}
-		phase, ok := nextAutoLaunchPhase(record)
-		if !ok {
-			continue
-		}
-		var cmd tea.Cmd
-		m, cmd = m.prepareAutoFlowPhaseLaunchForRecord(record, phase)
-		if cmd != nil {
-			cmds = append(cmds, cmd)
-		}
-	}
-	return m, batchNonNil(cmds...)
-}
-
-type deferredAutoFlowLaunchKey struct {
-	FlowID  string
-	PhaseID string
-}
-
-type suppressedAutoFlowLaunchKey struct {
-	FlowID   string
-	PhaseID  string
-	LaunchID string
-}
-
-func (m Model) deferAutoFlowPhaseLaunch(flowID, phaseID string) Model {
-	key, ok := newDeferredAutoFlowLaunchKey(flowID, phaseID)
-	if !ok {
-		return m
-	}
-	if m.deferredAutoFlowLaunches == nil {
-		m.deferredAutoFlowLaunches = make(map[deferredAutoFlowLaunchKey]struct{})
-	}
-	m.deferredAutoFlowLaunches[key] = struct{}{}
-	return m
-}
-
-func (m Model) suppressAutoFlowPhaseLaunch(flowID, phaseID, launchID string) Model {
-	key, ok := newSuppressedAutoFlowLaunchKey(flowID, phaseID, launchID)
-	if !ok {
-		return m
-	}
-	if m.suppressedAutoFlowLaunches == nil {
-		m.suppressedAutoFlowLaunches = make(map[suppressedAutoFlowLaunchKey]struct{})
-	}
-	m.suppressedAutoFlowLaunches[key] = struct{}{}
-	deferredKey, ok := newDeferredAutoFlowLaunchKey(flowID, phaseID)
-	if ok {
-		m = m.clearDeferredAutoFlowLaunch(deferredKey)
-	}
-	return m
-}
-
-func newSuppressedAutoFlowLaunchKey(flowID, phaseID, launchID string) (suppressedAutoFlowLaunchKey, bool) {
-	phaseID = artifacts.NormalizePhaseID(phaseID)
-	launchID = strings.TrimSpace(launchID)
-	if flowID == "" || phaseID == "" {
-		return suppressedAutoFlowLaunchKey{}, false
-	}
-	return suppressedAutoFlowLaunchKey{FlowID: flowID, PhaseID: phaseID, LaunchID: launchID}, true
-}
-
-func suppressedAutoFlowLaunchKeyForPhase(flowID string, phase flowstore.FlowPhase) (suppressedAutoFlowLaunchKey, bool) {
-	return newSuppressedAutoFlowLaunchKey(flowID, phase.PhaseID, flowstore.LatestPhaseLaunchID(phase))
-}
-
-func (m Model) isAutoFlowLaunchSuppressed(flowID string, phase flowstore.FlowPhase) bool {
-	key, ok := suppressedAutoFlowLaunchKeyForPhase(flowID, phase)
-	if !ok {
-		return false
-	}
-	_, suppressed := m.suppressedAutoFlowLaunches[key]
-	return suppressed
-}
-
-func (m Model) clearSuppressedAutoFlowLaunch(flowID string, phase flowstore.FlowPhase) Model {
-	key, ok := suppressedAutoFlowLaunchKeyForPhase(flowID, phase)
-	if !ok || len(m.suppressedAutoFlowLaunches) == 0 {
-		return m
-	}
-	delete(m.suppressedAutoFlowLaunches, key)
-	if len(m.suppressedAutoFlowLaunches) == 0 {
-		m.suppressedAutoFlowLaunches = nil
-	}
-	return m
-}
-
-func (m Model) clearResolvedSuppressedAutoFlowLaunches(record flowstore.FlowRecord) Model {
-	if len(m.suppressedAutoFlowLaunches) == 0 || record.FlowID == "" {
-		return m
-	}
-	for _, phase := range record.Phases {
-		if phase.Status == flowstore.PhaseRunning || phase.Status == flowstore.PhaseCompleted {
-			continue
-		}
-		m = m.clearSuppressedAutoFlowLaunch(record.FlowID, phase)
-	}
-	return m
-}
-
-func (m Model) clearDeferredAutoFlowLaunch(key deferredAutoFlowLaunchKey) Model {
-	if len(m.deferredAutoFlowLaunches) == 0 {
-		return m
-	}
-	delete(m.deferredAutoFlowLaunches, key)
-	if len(m.deferredAutoFlowLaunches) == 0 {
-		m.deferredAutoFlowLaunches = nil
-	}
-	return m
-}
-
-func newDeferredAutoFlowLaunchKey(flowID, phaseID string) (deferredAutoFlowLaunchKey, bool) {
-	phaseID = artifacts.NormalizePhaseID(phaseID)
-	if flowID == "" || phaseID == "" {
-		return deferredAutoFlowLaunchKey{}, false
-	}
-	return deferredAutoFlowLaunchKey{FlowID: flowID, PhaseID: phaseID}, true
-}
-
-func (m Model) prepareAutoFlowPhaseLaunchForRecord(record flowstore.FlowRecord, phase flowstore.FlowPhase) (Model, tea.Cmd) {
-	target, ok, next := m.flowPhaseLaunchTarget(FlowPhaseLaunchRequest{
-		Record:     record,
-		Phase:      phase,
-		AutoLaunch: true,
-		Headless:   m.flowHeadless,
-	})
-	m = next
-	if !ok {
-		return m, nil
-	}
-	return m, next.prepareFlowPhaseLaunch(target)
-}
-
-func (m Model) prepareDeferredAutoFlowPhaseLaunches() (Model, tea.Cmd) {
-	var cmds []tea.Cmd
-	for key := range m.deferredAutoFlowLaunches {
-		delete(m.deferredAutoFlowLaunches, key)
-		record, ok := m.flowByID(key.FlowID)
-		if !ok {
-			continue
-		}
-		if !record.AutoMode {
-			continue
-		}
-		sourcePhase, sourcePhaseOK := flowRecordPhaseByID(record, key.PhaseID)
-		sourceLaunchID := flowstore.LatestPhaseLaunchID(sourcePhase)
-		if sourcePhaseOK && m.hasRunningFlowEmbeddedTerminalForPhaseLaunch(key.FlowID, key.PhaseID, sourceLaunchID) {
-			m.deferredAutoFlowLaunches[key] = struct{}{}
-			continue
-		}
-		if sourcePhaseOK && m.hasAutoClosingFlowEmbeddedTerminalForPhaseLaunch(key.FlowID, key.PhaseID, sourceLaunchID) {
-			m.deferredAutoFlowLaunches[key] = struct{}{}
-			continue
-		}
-		if sourcePhaseOK && m.hasFlowEmbeddedTerminalForPhaseLaunch(key.FlowID, key.PhaseID, sourceLaunchID) {
-			continue
-		}
-		if phase, ok := nextAutoLaunchPhase(record); ok {
-			var cmd tea.Cmd
-			m, cmd = m.prepareAutoFlowPhaseLaunchForRecord(record, phase)
-			if cmd != nil {
-				cmds = append(cmds, cmd)
-			}
-		}
-	}
-	if len(m.deferredAutoFlowLaunches) == 0 {
-		m.deferredAutoFlowLaunches = nil
-	}
-	return m, batchNonNil(cmds...)
-}
-
 func (m Model) handleResetSelectedFlowPhase() (tea.Model, tea.Cmd) {
 	record, phase, repoPath, ok := m.selectedFlowPhaseResetTarget()
 	if !ok {
@@ -1537,23 +1325,6 @@ func (m Model) handleResetSelectedFlowPhase() (tea.Model, tea.Cmd) {
 		}
 	})
 	return m, nil
-}
-
-type flowPhaseLaunchTarget struct {
-	FlowPhaseLaunchPreparedRequest
-}
-
-func (m Model) selectedFlowNextLaunchTarget() (flowPhaseLaunchTarget, bool, Model) {
-	record, phase, ok := m.selectedFlowNextLaunchablePhase()
-	if !ok {
-		m = m.setStatus(statusOther, "No launchable Flow phase")
-		return flowPhaseLaunchTarget{}, false, m
-	}
-	return m.flowPhaseLaunchTarget(FlowPhaseLaunchRequest{
-		Record:   record,
-		Phase:    phase,
-		Headless: m.flowHeadless,
-	})
 }
 
 func (m Model) selectedFlowPhaseResetTarget() (flowstore.FlowRecord, flowstore.FlowPhase, string, bool) {
@@ -1693,35 +1464,6 @@ func (m Model) resetFlowPhaseCmd(repoPath, flowID, phaseID string) tea.Cmd {
 			Flow:     flow,
 		}
 	}
-}
-
-func (m Model) flowPhaseLaunchTarget(req FlowPhaseLaunchRequest) (flowPhaseLaunchTarget, bool, Model) {
-	prepared, err := m.flowPhaseLauncher().Preflight(req)
-	if err != nil {
-		m = m.setStatus(statusOther, err.Error())
-		return flowPhaseLaunchTarget{}, false, m
-	}
-	return flowPhaseLaunchTarget{FlowPhaseLaunchPreparedRequest: prepared}, true, m
-}
-
-func (m Model) prepareFlowPhaseLaunch(target flowPhaseLaunchTarget) tea.Cmd {
-	return func() tea.Msg {
-		result, err := m.flowPhaseLauncher().Prepare(target.FlowPhaseLaunchPreparedRequest)
-		if err != nil {
-			return ActionFailedMsg{RepoPath: target.RepoPath, Err: err.Error()}
-		}
-		if result.Skipped {
-			return nil
-		}
-		return m.flowPhaseLaunchMessage(result)
-	}
-}
-
-func (m Model) flowPhaseLaunchMessage(result FlowPhaseLaunchResult) tea.Msg {
-	if result.Route == FlowPhaseLaunchEmbedded {
-		return FlowEmbeddedLaunchRequestedMsg{LaunchContext: result.Context}
-	}
-	return PlanLaunchRequestedMsg{LaunchContext: result.Context}
 }
 
 func (m Model) handleResumeSession() (tea.Model, tea.Cmd) {
